move aggregation_functions to interface package
