+++ openfl/federated/plan/plan.py
@@ -12,8 +12,8 @@ from yaml import dump
 from yaml import safe_load
 from yaml import SafeDumper
 
-from openfl.component.aggregation_functions import AggregationFunction
-from openfl.component.aggregation_functions import WeightedAverage
+from openfl.interface.aggregation_functions import AggregationFunction
+from openfl.interface.aggregation_functions import WeightedAverage
 from openfl.component.assigner.custom_assigner import Assigner
 from openfl.interface.cli_helper import WORKSPACE
 from openfl.transport import AggregatorGRPCClient
@@ -302,7 +302,7 @@ class Plan:
                 aggregation_type = Plan.build(**aggregation_type)
                 if not isinstance(aggregation_type, AggregationFunction):
                     raise NotImplementedError(f'''{task} task aggregation type does not implement an interface:
-    openfl.component.aggregation_functions.AggregationFunction
+    openfl.interface.aggregation_functions.AggregationFunction
     ''')
             tasks[task]['aggregation_type'] = aggregation_type
         return tasks
