+++ openfl/interface/aggregation_functions/__init__.py
