+++ openfl/interface/aggregation_functions/core/__init__.py
