+++ openfl/interface/aggregation_functions/core/interface.py
