+++ openfl/interface/aggregation_functions/geometric_median.py
