+++ openfl/interface/aggregation_functions/weighted_average.py
