+++ openfl/interface/aggregation_functions/yogi_adaptive_aggregation.py
