+++ setup.py
@@ -89,9 +89,9 @@ setup(
     packages=[
         'openfl',
         'openfl.component',
-        'openfl.component.aggregation_functions',
-        'openfl.component.aggregation_functions.core',
-        'openfl.component.aggregation_functions.experimental',
+        'openfl.interface.aggregation_functions',
+        'openfl.interface.aggregation_functions.core',
+        'openfl.interface.aggregation_functions.experimental',
         'openfl.component.aggregator',
         'openfl.component.assigner',
         'openfl.component.ca',
